Move helper to class contained in module
